fix: release private response accumulator state safely
