+++ oxlint.config.ts
@@ -164,7 +164,7 @@ module.exports = defineConfig({
         'src/lib/ChatCompletionStream.ts',
         'src/lib/ResponsesParser.ts',
         'src/lib/parser.ts',
-        'src/internal/responses/ResponseAccumulator.ts',
+        'src/lib/responses/ResponseAccumulator.ts',
         'src/lib/responses/ResponseInputItems.ts',
         'src/lib/responses/ResponseStream.ts',
         'src/lib/transform.ts',
@@ -659,7 +659,7 @@ module.exports = defineConfig({
         'src/lib/ChatCompletionStream.ts',
         'src/lib/ResponsesParser.ts',
         'src/lib/chatCompletionUtils.ts',
-        'src/internal/responses/ResponseAccumulator.ts',
+        'src/lib/responses/ResponseAccumulator.ts',
         'src/lib/responses/ResponseStream.ts',
       ],
       rules: {
@@ -877,7 +877,6 @@ module.exports = defineConfig({
         'src/lib/ResponsesParser.ts',
         'src/lib/RunnableFunction.ts',
         'src/lib/parser.ts',
-        'src/internal/responses/ResponseAccumulator.ts',
         'src/lib/responses/ResponseAccumulator.ts',
         'src/lib/responses/ResponseStream.ts',
         'src/lib/transform.ts',
@@ -966,7 +965,7 @@ module.exports = defineConfig({
         'src/internal/ws.ts',
         'src/lib/AssistantStream.ts',
         'src/lib/EventStream.ts',
-        'src/internal/responses/ResponseAccumulator.ts',
+        'src/lib/responses/ResponseAccumulator.ts',
         'src/lib/responses/ResponseInputItems.ts',
         'src/lib/transform.ts',
         'src/providers/bedrock/aws.ts',
@@ -1119,7 +1118,7 @@ module.exports = defineConfig({
         'src/lib/ChatCompletionStreamingRunner.ts',
         'src/lib/ResponsesParser.ts',
         'src/lib/parser.ts',
-        'src/internal/responses/ResponseAccumulator.ts',
+        'src/lib/responses/ResponseAccumulator.ts',
         'src/lib/responses/ResponseStream.ts',
       ],
       rules: {
@@ -1316,7 +1315,6 @@ module.exports = defineConfig({
         'src/lib/Util.ts',
         'src/lib/chatCompletionUtils.ts',
         'src/lib/responses/EventTypes.ts',
-        'src/internal/responses/ResponseAccumulator.ts',
         'src/lib/responses/ResponseAccumulator.ts',
         'src/lib/responses/ResponseInputItems.ts',
         'src/lib/responses/ResponseStream.ts',
@@ -1372,7 +1370,7 @@ module.exports = defineConfig({
         'src/lib/AbstractChatCompletionRunner.ts',
         'src/lib/AssistantStream.ts',
         'src/lib/ChatCompletionStream.ts',
-        'src/internal/responses/ResponseAccumulator.ts',
+        'src/lib/responses/ResponseAccumulator.ts',
         'src/lib/responses/ResponseInputItems.ts',
         'src/lib/transform.ts',
       ],
@@ -1396,7 +1394,7 @@ module.exports = defineConfig({
         'src/core/streaming.ts',
         'src/internal/to-file.ts',
         'src/lib/EventStream.ts',
-        'src/internal/responses/ResponseAccumulator.ts',
+        'src/lib/responses/ResponseAccumulator.ts',
         'src/realtime/internal-base.ts',
         'src/realtime/websocket.ts',
         'src/realtime/ws.ts',
